Miglioramento RF6 e terminazione della sessione chat visibile
