--- ALEC_Chatbot/last_report.html
@@ -1,95 +1,0 @@
-
-    <!DOCTYPE html>
-    <html lang='it'>
-    <head>
-        <meta charset='UTF-8'>
-        <title>Report finale ALEC</title>
-        <style>
-            body { font-family: sans-serif; max-width: 800px; margin: auto; padding: 20px; }
-            h1 { color: darkgreen; }
-            hr { margin: 1em 0; }
-            p, li { line-height: 1.5em; }
-            ul { margin-left: 1.2em; }
-        </style>
-    </head>
-    <body>
-        <h1>📘 Report finale di sessione</h1>
-        <hr>
-        <!DOCTYPE html>
-<html lang="it">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Report Finale di Sessione</title>
-    <style>
-        body { font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif; line-height: 1.6; color: #333; margin: 20px; background-color: #f8f9fa; }
-        h1 { color: #2c3e50; text-align: center; margin-bottom: 30px; }
-        h2 { color: #34495e; border-bottom: 2px solid #3498db; padding-bottom: 8px; margin-top: 30px; }
-        h3 { color: #2980b9; margin-top: 20px; }
-        section { margin-bottom: 30px; padding: 20px; border: 1px solid #e0e0e0; border-radius: 8px; background-color: #ffffff; box-shadow: 0 2px 4px rgba(0,0,0,0.05); }
-        ul { list-style-type: disc; margin-left: 25px; padding: 0; }
-        ol { list-style-type: decimal; margin-left: 25px; padding: 0; }
-        li { margin-bottom: 8px; }
-        .student-summary { margin-top: 20px; border-left: 4px solid #3498db; padding-left: 15px; background-color: #f2faff; border-radius: 4px; padding-top: 10px; padding-bottom: 10px; }
-        .highlight { font-weight: bold; color: #e74c3c; } /* Emphasis for key points */
-        p { margin-bottom: 10px; }
-        code { background-color: #eee; padding: 2px 4px; border-radius: 3px; font-family: 'Consolas', 'Courier New', monospace; }
-    </style>
-</head>
-<body>
-    <h1>Report Finale di Sessione</h1>
-
-    <section>
-        <h2>Docente</h2>
-        <p>
-            La sessione ha evidenziato un <span class="highlight">coinvolgimento costante da parte di entrambi gli studenti</span>, con un impegno medio di due turni di intervento ciascuno. Questo indica una buona proattività nel partecipare alle discussioni.
-        </p>
-        <p>
-            L'andamento generale ha mostrato una <span class="highlight">variabilità nella qualità e nella pertinenza delle risposte</span>. Mentre un partecipante ha dimostrato eccellente qualità quando pertinente, l'altro ha indicato la necessità di maggiore approfondimento.
-        </p>
-        <p>
-            Le fonti più frequentemente consultate hanno ruotato attorno agli <span class="highlight">"algoritmi di ricerca locale"</span>, con una risorsa specifica, <code>algoritmi_ricerca_locale_GA_opt.pdf</code>, utilizzata da entrambi gli studenti. Sono state esplorate anche aree come la ricerca con avversari e la ricerca informata.
-        </p>
-        <p>
-            Le difficoltà comuni riscontrate riguardano la <span class="highlight">trasformazione della consultazione delle fonti in interventi di alta qualità e costantemente pertinenti</span>. È emersa l'importanza di non solo individuare le informazioni, ma di assimilarle criticamente e applicarle in modo efficace.
-        </p>
-    </section>
-
-    <section>
-        <h2>Studenti</h2>
-
-        <div class="student-summary">
-            <h3>Davide</h3>
-            <ul>
-                <li><strong>Punti di forza:</strong> Grande <span class="highlight">proattività e apertura alla ricerca</span>, dimostrata dalla partecipazione attiva con due turni e dalla consultazione di una varietà di fonti. Questo spirito di esplorazione è una base solida.</li>
-                <li><strong>Aree di miglioramento:</strong> È cruciale focalizzarsi sulla <span class="highlight">qualità e la pertinenza specifica delle risposte</span>. La qualità media di 0.00 suggerisce che, pur essendoci un contributo, la profondità e l'accuratezza possono essere significativamente migliorate.</li>
-                <li><strong>Suggerimenti di studio personalizzati:</strong>
-                    <ol>
-                        <li>Prima di fornire una risposta, prenditi un momento per <span class="highlight">riflettere e strutturare il tuo pensiero</span>, assicurandoti che sia direttamente allineato con l'obiettivo della discussione.</li>
-                        <li>Concentrati sull'<span class="highlight">analisi critica e sull'applicazione pratica</span> dei concetti dalle fonti, piuttosto che sulla mera individuazione delle stesse. Cerca di capire il "perché" dietro le informazioni.</li>
-                        <li>Non esitare a chiedere <span class="highlight">feedback specifici</span> sui tuoi interventi. Questo ti aiuterà a identificare le aree precise in cui migliorare la tua comprensione e la chiarezza espositiva.</li>
-                    </ol>
-                </li>
-            </ul>
-        </div>
-
-        <div class="student-summary">
-            <h3>Marcello</h3>
-            <ul>
-                <li><strong>Punti di forza:</strong> <span class="highlight">Ottima qualità nelle risposte pertinenti</span> (qualità media 1.00), il che evidenzia una solida comprensione e capacità di sintesi quando il contributo è mirato. L'approccio focalizzato sulle fonti di "ricerca locale" è risultato efficace.</li>
-                <li><strong>Aree di miglioramento:</strong> Mantenere e <span class="highlight">aumentare la frequenza di risposte pertinenti e di alta qualità</span>. L'obiettivo è estendere l'eccellente livello di qualità già dimostrato a un maggior numero di interventi, massimizzando il tuo impatto nella discussione.</li>
-                <li><strong>Suggerimenti di studio personalizzati:</strong>
-                    <ol>
-                        <li>Continua a <span class="highlight">capitalizzare sulla tua profondità di comprensione</span> negli argomenti di ricerca locale. Potresti esplorare sfumature più complesse o applicazioni avanzate di questi concetti.</li>
-                        <li>Prepara attivamente il terreno per i tuoi interventi. Puoi <span class="highlight">formulare in anticipo possibili domande o osservazioni</span> basate sulle tue letture, così da essere pronto a contribuire in vari momenti.</li>
-                        <li>Cerca di <span class="highlight">collegare i concetti</span> in cui sei forte con altri temi o discipline. Questo ti permetterà di dimostrare versatilità e ampliare il raggio di pertinenza dei tuoi contributi.</li>
-                    </ol>
-                </li>
-            </ul>
-        </div>
-    </section>
-</body>
-</html>
-    </body>
-    </html>
-
