+++ ALEC_Chatbot/requirements.txt
@@ -4,3 +4,4 @@ protobuf~=5.29.5
 numpy~=2.3.3
 sentence-transformers~=5.1.1
 fastapi~=0.119.0
+uvicorn~=0.37.0
