+++ ALEC_Chatbot/src/chatbot/app/templates/homepage.html
@@ -1,0 +1,95 @@
+<!DOCTYPE html>
+        <html lang="it">
+        <head>
+            <meta charset="UTF-8">
+            <title>ALEC - Peer Tutor Simulation</title>
+            <style>
+                body { font-family: sans-serif; max-width: 700px; margin: auto; padding: 20px; }
+                textarea { width: 100%; height: 80px; }
+                .log { border: 1px solid #ccc; padding: 10px; height: 300px; overflow-y: auto; margin-bottom: 1em; }
+                .alec { color: darkgreen; }
+                .marcello { color: blue; }
+                .davide { color: darkred; }
+            </style>
+        </head>
+        <body>
+            <h2>Simulazione PeerTutor (due studenti)</h2>
+            <div class="log" id="log"></div>
+
+            <select id="speaker">
+                <option value="Davide">Davide</option>
+                <option value="Marcello">Marcello</option>
+            </select>
+            <textarea id="msg" placeholder="Scrivi un messaggio..."></textarea><br/>
+            <button onclick="send()">Invia</button>
+
+           <script>
+                async function send() {
+                    const speaker = document.getElementById('speaker').value;
+                    const message = document.getElementById('msg').value.trim();
+                    const log = document.getElementById('log');
+                    if (!message) return;
+
+                    // Mostra subito il messaggio dell'utente
+                    log.innerHTML += `<div class="${speaker.toLowerCase()}"><b>${speaker}:</b> ${message}</div>`;
+                    document.getElementById('msg').value = '';
+
+                    // --- Intercetta comando /exit ---
+                    const isExit = message.toLowerCase() === "/exit";
+                    if (isExit) {
+                        // Blocca subito UI
+                        document.getElementById("msg").disabled = true;
+                        const btn = document.querySelector("button[onclick='send()']");
+                        btn.disabled = true;
+                        btn.innerText = "Sessione chiusa 🔒";
+                        btn.style.backgroundColor = "#999";
+                        btn.style.cursor = "not-allowed";
+                        log.innerHTML += `<div><i>⏳ Generazione del report in corso...</i></div>`;
+                    }
+
+                    try {
+                        const resp = await fetch('/message', {
+                            method: 'POST',
+                            headers: {'Content-Type': 'application/json'},
+                            body: JSON.stringify({speaker, message})
+                        });
+
+                        const data = await resp.json();
+
+                        // Mostra la risposta di ALEC
+                        if (data.message) {
+                            log.innerHTML += `<div class="alec"><b>ALEC → ${speaker}:</b> ${data.message}</div>`;
+                        } else {
+                            log.innerHTML += `<div><i>Errore:</i> ${JSON.stringify(data)}</div>`;
+                        }
+
+                        // --- Gestione flag session_closed dal server ---
+                        if (data.session_closed === true && !isExit) {
+                            document.getElementById("msg").disabled = true;
+                            const btn = document.querySelector("button[onclick='send()']");
+                            btn.disabled = true;
+                            btn.innerText = "Sessione chiusa 🔒";
+                            btn.style.backgroundColor = "#999";
+                            btn.style.cursor = "not-allowed";
+                            log.innerHTML += `<div><i>🔒 Sessione terminata.</i></div>`;
+                        }
+
+                    } catch (err) {
+                        // In caso di errore di rete, riattiva il bottone se non era un /exit
+                        log.innerHTML += `<div><i>Errore di rete:</i> ${err}</div>`;
+                        if (!isExit) {
+                            document.getElementById("msg").disabled = false;
+                            const btn = document.querySelector("button[onclick='send()']");
+                            btn.disabled = false;
+                            btn.innerText = "Invia";
+                            btn.style.backgroundColor = "";
+                            btn.style.cursor = "pointer";
+                        }
+                    }
+
+                    // Scorri in basso
+                    log.scrollTop = log.scrollHeight;
+                }
+           </script>
+        </body>
+</html>
